apps/api/requirements.txt
# Django and REST Framework
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# AI Integration
google-generativeai>=0.4,<1.0

# Utils
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0

# Development
gunicorn>=21.0,<22.0
